backend/requirements.txt
# ─── API Server ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── AI / ML Core ───────────────────────────────────────────────────────────
# PyTorch — install the CUDA build separately if on GPU:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch==2.3.0
torchvision==0.18.0

# Meta SAM
git+https://github.com/facebookresearch/segment-anything.git

# HuggingFace (DINOv2)
transformers==4.41.0
huggingface-hub==0.23.0

# ─── Classical CV ───────────────────────────────────────────────────────────
opencv-python-headless==4.9.0.80

# ─── Numerical / Scientific ─────────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.0

# ─── Image I/O ──────────────────────────────────────────────────────────────
Pillow==10.3.0

# ─── Data Validation ────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Job Store ──────────────────────────────────────────────────────────────
redis==5.0.4

# ─── Logging ────────────────────────────────────────────────────────────────
structlog==24.1.0

# ─── Testing ────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
asgi-lifespan==2.1.0

# ─── Dev Utilities ──────────────────────────────────────────────────────────
python-dotenv==1.0.1